+++ CHANGELOG.md
@@ -3,13 +3,13 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.54.1] - 2018-06-20
 ---------------------
 ##### Fixed
 - In a form containing a group with a single child repeat (and no other repeat sibling questions), fails to load a record where that group is empty. When creating the first repeat after load an exception occurs.
+- Output inside a group that is irrelevant upon loading does not get evaluated the group becomes relevant.
 
-
-[4.54.0] - 2018-05-18
+[4.54.0] - 2018-06-18
 ---------------------
 ##### Added
 - URL widget